fix(StandardTablePagination): remove unused capitalLabel variable
