--- Clients/src/presentation/components/Table/EvaluationTable/TableHead/index.tsx
@@ -1,131 +1,0 @@
-import { TableHead, TableRow, TableCell, Box, Typography, useTheme } from "@mui/material";
-import { ChevronsUpDown, ChevronUp, ChevronDown } from "lucide-react";
-import singleTheme from "../../../../themes/v1SingleTheme";
-import { palette } from "../../../../themes/palette";
-
-// Column width definitions for consistent spacing
-const columnWidths: Record<string, string> = {
-  "EXPERIMENT NAME": "20%",
-  "MODEL": "12%",
-  "JUDGE/SCORER": "16%",
-  "# PROMPTS": "8%",
-  "DATASET": "14%",
-  "DATE": "16%",
-  "ACTION": "60px",
-};
-
-// Map column labels to sortable field keys
-const columnSortKeys: Record<string, string> = {
-  "EXPERIMENT NAME": "name",
-  "MODEL": "model",
-  "JUDGE/SCORER": "judge",
-  "# PROMPTS": "prompts",
-  "DATASET": "dataset",
-  "DATE": "date",
-};
-
-export type SortDirection = "asc" | "desc" | null;
-export type SortConfig = {
-  key: string;
-  direction: SortDirection;
-};
-
-interface TableHeaderProps {
-  columns: string[];
-  sortConfig?: SortConfig;
-  onSort?: (columnKey: string) => void;
-}
-
-const TableHeader = ({ columns, sortConfig, onSort }: TableHeaderProps) => {
-  const theme = useTheme();
-  // Columns that should be center-aligned
-  const centerAlignedColumns = ["MODEL", "JUDGE/SCORER", "# PROMPTS", "DATASET", "DATE", "ACTION"];
-
-  return (
-    <TableHead
-      sx={{
-        backgroundColor:
-          singleTheme.tableStyles.primary.header.backgroundColors,
-      }}
-    >
-      <TableRow sx={singleTheme.tableStyles.primary.header.row}>
-        {columns.map((column, index) => {
-          const isActionColumn = column === "ACTION" || column === "Actions";
-          const isCenterAligned = centerAlignedColumns.includes(column);
-          const isFirstColumn = index === 0;
-          const width = columnWidths[column];
-          const sortKey = columnSortKeys[column];
-          const isSortable = !!sortKey && onSort;
-          
-          return (
-            <TableCell
-              key={index}
-              sx={{
-                ...singleTheme.tableStyles.primary.header.cell,
-                width: width || "auto",
-                ...(isActionColumn
-                  ? {
-                      minWidth: "60px",
-                      maxWidth: "60px",
-                    }
-                  : {}),
-                ...(isSortable
-                  ? {
-                      cursor: "pointer",
-                      userSelect: "none",
-                      "&:hover": {
-                        backgroundColor: "rgba(0, 0, 0, 0.04)",
-                      },
-                    }
-                  : {}),
-              }}
-              onClick={() => isSortable && onSort(sortKey)}
-            >
-              <Box
-                sx={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: isFirstColumn ? "flex-start" : isCenterAligned ? "center" : "flex-start",
-                  gap: theme.spacing(2),
-                }}
-              >
-                <Typography
-                  variant="body2"
-                  sx={{
-                    fontWeight: 500,
-                    fontSize: "13px",
-                    color: sortConfig?.key === sortKey ? "primary.main" : "inherit",
-                  }}
-                >
-                  {column}
-                </Typography>
-                {isSortable && (
-                  <Box
-                    sx={{
-                      display: "flex",
-                      alignItems: "center",
-                      color: sortConfig?.key === sortKey ? palette.brand.primary : palette.text.disabled,
-                    }}
-                  >
-                    {sortConfig?.key === sortKey && sortConfig?.direction === "asc" && (
-                      <ChevronUp size={14} />
-                    )}
-                    {sortConfig?.key === sortKey && sortConfig?.direction === "desc" && (
-                      <ChevronDown size={14} />
-                    )}
-                    {sortConfig?.key !== sortKey && (
-                      <ChevronsUpDown size={14} />
-                    )}
-                  </Box>
-                )}
-              </Box>
-            </TableCell>
-          );
-        })}
-      </TableRow>
-    </TableHead>
-  );
-};
-
-export default TableHeader;
-
