+++ Clients/src/presentation/components/Table/EvaluationTable/index.tsx
@@ -10,7 +10,7 @@ import {
 } from "@mui/material";
 import { Suspense, lazy, useMemo, useState, useCallback, useEffect } from "react";
 import TablePaginationActions from "../../TablePagination";
-import TableHeader, { SortConfig } from "./TableHead";
+import StandardTableHead from "../StandardTableHead";
 import { EmptyState } from "../../EmptyState";
 import EmptyStateTip from "../../EmptyState/EmptyStateTip";
 import { ChevronsUpDown, FlaskConical, Play, Database, BarChart3 } from "lucide-react";
@@ -30,10 +30,41 @@ import {
   setPaginationRowCount,
 } from "../../../../application/utils/paginationStorage";
 import { IEvaluationTableProps, IEvaluationRow } from "../../../types/interfaces/i.table";
+import type { SortConfig, StandardColumn } from "../../../../domain/types/standardTable";
 
 const EvaluationTableBody = lazy(() => import("./TableBody"));
 
 const EVALUATION_SORTING_KEY = "verifywise_evaluation_sorting";
+
+// Map column labels to sortable field keys
+const columnSortKeys: Record<string, string> = {
+  "EXPERIMENT NAME": "name",
+  "MODEL": "model",
+  "JUDGE/SCORER": "judge",
+  "# PROMPTS": "prompts",
+  "DATASET": "dataset",
+  "DATE": "date",
+};
+
+// Column width definitions for consistent spacing
+const columnWidths: Record<string, string> = {
+  "EXPERIMENT NAME": "20%",
+  "MODEL": "12%",
+  "JUDGE/SCORER": "16%",
+  "# PROMPTS": "8%",
+  "DATASET": "14%",
+  "DATE": "16%",
+  "ACTION": "60px",
+};
+
+const toStandardColumns = (labels: string[]): StandardColumn[] =>
+  labels.map((label) => ({
+    id: columnSortKeys[label] || label.toLowerCase(),
+    label,
+    sortable: !!columnSortKeys[label],
+    width: columnWidths[label],
+    minWidth: label === "ACTION" ? "60px" : undefined,
+  }));
 
 const EvaluationTable: React.FC<IEvaluationTableProps> = ({
   columns,
@@ -47,6 +78,8 @@ const EvaluationTable: React.FC<IEvaluationTableProps> = ({
   onCopy,
   hidePagination = false,
 }) => {
+  const standardColumns = useMemo(() => toStandardColumns(columns), [columns]);
+
   const [rowsPerPage, setRowsPerPage] = useState(() =>
     getPaginationRowCount("evaluation", 10)
   );
@@ -57,7 +90,6 @@ const EvaluationTable: React.FC<IEvaluationTableProps> = ({
     if (saved) {
       try {
         const parsed = JSON.parse(saved);
-        // If saved state is empty, use the default
         if (!parsed.key || !parsed.direction) {
           return { key: "date", direction: "desc" };
         }
@@ -105,43 +137,35 @@ const EvaluationTable: React.FC<IEvaluationTableProps> = ({
           aValue = a.id.toLowerCase();
           bValue = b.id.toLowerCase();
           break;
-
         case "model":
           aValue = a.model.toLowerCase();
           bValue = b.model.toLowerCase();
           break;
-
         case "judge":
           aValue = (a.judge || "").toLowerCase();
           bValue = (b.judge || "").toLowerCase();
           break;
-
         case "prompts":
           aValue = a.prompts ?? 0;
           bValue = b.prompts ?? 0;
           break;
-
         case "dataset":
           aValue = a.dataset.toLowerCase();
           bValue = b.dataset.toLowerCase();
           break;
-
         case "date":
           aValue = a.date ? new Date(a.date).getTime() : 0;
           bValue = b.date ? new Date(b.date).getTime() : 0;
           break;
-
         default:
           return 0;
       }
 
-      // Handle string comparisons
       if (typeof aValue === "string" && typeof bValue === "string") {
         const comparison = aValue.localeCompare(bValue);
         return sortConfig.direction === "asc" ? comparison : -comparison;
       }
 
-      // Handle number comparisons
       if (aValue < bValue) return sortConfig.direction === "asc" ? -1 : 1;
       if (aValue > bValue) return sortConfig.direction === "asc" ? 1 : -1;
       return 0;
@@ -178,7 +202,11 @@ const EvaluationTable: React.FC<IEvaluationTableProps> = ({
       <TableContainer sx={{ mt: 10 }}>
         <Suspense fallback={<div>Loading...</div>}>
           <Table sx={{ ...singleTheme.tableStyles.primary.frame }}>
-            <TableHeader columns={columns} sortConfig={sortConfig} onSort={handleSort} />
+            <StandardTableHead
+              columns={standardColumns}
+              sortConfig={sortConfig}
+              onSort={handleSort}
+            />
             {sortedRows.length !== 0 ? (
               <>
                 <EvaluationTableBody
